Only use property editors via reflection from main code, and don't include the class accessing java.bean in slimJar
