+++ src/fitnesse/slim/converters/ConverterRegistry.java
@@ -2,6 +2,8 @@ package fitnesse.slim.converters;
 
 import fitnesse.slim.Converter;
 
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.lang.reflect.ParameterizedType;
 import java.util.*;
 
@@ -9,9 +11,19 @@ public class ConverterRegistry {
 
   private static final Map<Class<?>, Converter<?>> converters = new HashMap<>();
   private static Converter<Object> defaultConverter = new DefaultConverter();
+  private static Method javaBeansGetConverterMethod;
 
   static {
     addStandardConverters();
+
+    // java.beans is not available on Android, so we cannot use JavaBeansPropertyEditorConverterFactory
+    // directly but only via reflection
+    try {
+      Class<?> factory = Class.forName("fitnesse.slim.converters.beans.JavaBeansPropertyEditorConverterFactory");
+      javaBeansGetConverterMethod = factory.getMethod("getConverter", Class.class);
+    } catch (ClassNotFoundException e) {
+    } catch (NoSuchMethodException e) {
+    }
   }
 
   private ConverterRegistry() {
@@ -75,11 +87,10 @@ public class ConverterRegistry {
       return converterForInterface;
     }
 
-    //use property editor
-    PropertyEditor pe = PropertyEditorManager.findEditor(clazz);
-    if (pe != null && !"EnumEditor".equals(pe.getClass().getSimpleName())) {
-      // com.sun.beans.EnumEditor and sun.beans.EnumEditor seem to be used in different usages.
-      return new PropertyEditorConverter<>(pe);
+    //use java beans property editor (does not work on Android)
+    Converter<T> javaBeansConverter = getJavaBeansBasedConverter(clazz);
+    if (javaBeansConverter != null) {
+      return javaBeansConverter;
     }
 
     //for enum, use generic enum converter
@@ -103,6 +114,18 @@ public class ConverterRegistry {
 
     // last resort, see if there is a converter for Object
     return (Converter<T>) converters.get(Object.class);
+  }
+
+  @SuppressWarnings("unchecked")
+  private static <T> Converter<T> getJavaBeansBasedConverter(Class<? extends T> clazz) {
+    if (javaBeansGetConverterMethod != null) {
+      try {
+        return (Converter<T>) javaBeansGetConverterMethod.invoke(null, clazz);
+      } catch (IllegalAccessException e) {
+      } catch (InvocationTargetException e) {
+      }
+    }
+    return null;
   }
 
   protected static <T> Converter<T> getConverterForInterface(Class<?> clazz) {
